+++ includes/database-access.php
@@ -117,7 +117,13 @@ class Ekc_Database_Access {
 		global $wpdb;
 		$row = $wpdb->get_row( $wpdb->prepare( 
 			"
-			SELECT tournament_id, code_name, name, tournament_date, team_size, max_teams, is_wait_list_enabled, is_player_names_required, is_auto_backup_enabled, tournament_system, elimination_rounds, elimination_max_points_per_round, swiss_system_rounds, swiss_system_max_points_per_round, swiss_system_additional_rounds, swiss_system_slide_match_rounds, swiss_system_round_time, swiss_system_start_pitch, shareable_link_url_prefix, shareable_link_email_text, shareable_link_sender_email
+			SELECT tournament_id, code_name, name, tournament_date,
+			       team_size, max_teams, is_wait_list_enabled, 
+				   is_player_names_required, is_auto_backup_enabled, 
+				   tournament_system, elimination_rounds, elimination_max_points_per_round, 
+				   swiss_system_rounds, swiss_system_max_points_per_round, swiss_system_additional_rounds, 
+				   swiss_system_slide_match_rounds, swiss_system_round_time, swiss_system_start_pitch, 
+				   shareable_link_url_prefix, shareable_link_email_text, shareable_link_sender_email
 			FROM   {$wpdb->prefix}ekc_tournament
 			WHERE  tournament_id = %d
 			",
@@ -135,7 +141,13 @@ class Ekc_Database_Access {
 		global $wpdb;
 		$row = $wpdb->get_row( $wpdb->prepare( 
 			"
-			SELECT tournament_id, code_name, name, tournament_date, team_size, max_teams, is_wait_list_enabled, is_player_names_required, is_auto_backup_enabled, tournament_system, elimination_rounds, elimination_max_points_per_round, swiss_system_rounds, swiss_system_max_points_per_round, swiss_system_additional_rounds, swiss_system_slide_match_rounds, swiss_system_round_time, swiss_system_start_pitch, shareable_link_url_prefix, shareable_link_email_text, shareable_link_sender_email
+			SELECT tournament_id, code_name, name, tournament_date, 
+			       team_size, max_teams, is_wait_list_enabled, 
+				   is_player_names_required, is_auto_backup_enabled, 
+				   tournament_system, elimination_rounds, elimination_max_points_per_round, 
+				   swiss_system_rounds, swiss_system_max_points_per_round, swiss_system_additional_rounds, 
+				   swiss_system_slide_match_rounds, swiss_system_round_time, swiss_system_start_pitch, 
+				   shareable_link_url_prefix, shareable_link_email_text, shareable_link_sender_email
 			FROM   {$wpdb->prefix}ekc_tournament
 			WHERE  code_name = %s
 			",
@@ -325,7 +337,13 @@ class Ekc_Database_Access {
 
 		$results = $wpdb->get_results( $wpdb->prepare(
 			"
-			SELECT t.team_id, t.tournament_id, t.name, LOWER(t.country) as country, t.is_active, t.email, t.phone, t.registration_date, t.camping_count, t.breakfast_count, t.is_on_wait_list, t.registration_order, t.seeding_score, t.initial_score, t.virtual_rank, t.shareable_link_id
+			SELECT t.team_id, t.tournament_id, t.name, 
+			       LOWER(t.country) as country, t.club,
+				   t.is_active, t.email, t.phone, t.registration_date, 
+				   t.camping_count, t.breakfast_count, 
+				   t.is_on_wait_list, t.registration_order, 
+				   t.seeding_score, t.initial_score, t.virtual_rank, 
+				   t.shareable_link_id
 			FROM   {$wpdb->prefix}ekc_team t
 			INNER JOIN (
 				SELECT x.team_id
@@ -358,15 +376,16 @@ class Ekc_Database_Access {
 		$team = new Ekc_Team();
 		$team->set_team_id( $row->team_id );
 		$team->set_tournament_id( $row->tournament_id );
-		$team->set_name( strval( $row->name ) );
-		$team->set_country( strval( $row->country) );
-		$team->set_active( boolval( $row->is_active ) );
-		$team->set_email( strval( $row->email ) );
-		$team->set_phone( strval( $row->phone ) );
-		$team->set_registration_date( strval( $row->registration_date ) );
+		$team->set_name( strval( $row->name ));
+		$team->set_country( strval( $row->country ));
+		$team->set_club( strval( $row->club ));
+		$team->set_active( boolval( $row->is_active ));
+		$team->set_email( strval( $row->email ));
+		$team->set_phone( strval( $row->phone ));
+		$team->set_registration_date( strval( $row->registration_date ));
 		$team->set_camping_count( $row->camping_count );
 		$team->set_breakfast_count( $row->breakfast_count );
-		$team->set_on_wait_list( boolval( $row->is_on_wait_list ) );
+		$team->set_on_wait_list( boolval( $row->is_on_wait_list ));
 		$team->set_registration_order( $row->registration_order );
 		$team->set_seeding_score( $row->seeding_score );
 		$team->set_initial_score( $row->initial_score );
@@ -406,7 +425,13 @@ class Ekc_Database_Access {
 
 		$results = $wpdb->get_results( $wpdb->prepare(
 			"
-			SELECT t.team_id, t.tournament_id, t.name, LOWER(t.country) as country, t.is_active, t.email, t.phone, t.registration_date, t.camping_count, t.breakfast_count, t.is_on_wait_list, t.registration_order, t.seeding_score, t.initial_score, t.virtual_rank, t.shareable_link_id 
+			SELECT t.team_id, t.tournament_id, t.name, 
+			       LOWER(t.country) as country, t.club,
+				   t.is_active, t.email, t.phone, t.registration_date, 
+				   t.camping_count, t.breakfast_count, 
+				   t.is_on_wait_list, t.registration_order, 
+				   t.seeding_score, t.initial_score, t.virtual_rank, 
+				   t.shareable_link_id 
 			FROM   {$wpdb->prefix}ekc_team t
 			WHERE  COALESCE(t.is_active, 0) = 1
 			AND    t.tournament_id = %d
@@ -433,7 +458,13 @@ class Ekc_Database_Access {
 		global $wpdb;
 		$results = $wpdb->get_results( $wpdb->prepare(
 			"
-			SELECT t.team_id, t.tournament_id, t.name, LOWER(t.country) as country, t.is_active, t.email, t.phone, t.registration_date, t.camping_count, t.breakfast_count, t.is_on_wait_list, t.registration_order, t.seeding_score, t.initial_score, t.virtual_rank, t.shareable_link_id
+			SELECT t.team_id, t.tournament_id, t.name, 
+			       LOWER(t.country) as country, t.club, 
+				   t.is_active, t.email, t.phone, t.registration_date, 
+				   t.camping_count, t.breakfast_count, 
+				   t.is_on_wait_list, t.registration_order, 
+				   t.seeding_score, t.initial_score, t.virtual_rank, 
+				   t.shareable_link_id
 			FROM   {$wpdb->prefix}ekc_team t
 			WHERE  t.tournament_id = %d
 			ORDER BY t.team_id ASC",
@@ -458,14 +489,19 @@ class Ekc_Database_Access {
 		$sql_filter = $this->create_team_table_filter( $filter, 't' );
 		$results = $wpdb->get_results( $wpdb->prepare(
 			"
-			SELECT t.team_id, t.name, LOWER(t.country) as country, case when t.is_active = 1 then 'yes' else 'no' end as is_active, t.email, t.phone, t.registration_date, t.registration_order, t.camping_count, t.breakfast_count, t.seeding_score, case when t.is_on_wait_list = 1 then 'yes' else 'no' end as is_on_wait_list,
-                               GROUP_CONCAT( CONCAT(p.first_name, ' ', p.last_name, ' (', p.country, ')') SEPARATOR ', ') as players
+			SELECT t.team_id, t.name, 
+			       LOWER(t.country) as country, t.club,
+				   case when t.is_active = 1 then 'yes' else 'no' end as is_active, 
+				   t.email, t.phone, t.registration_date, t.registration_order, 
+				   t.seeding_score, 
+				   case when t.is_on_wait_list = 1 then 'yes' else 'no' end as is_on_wait_list,
+                   GROUP_CONCAT( CONCAT(p.first_name, ' ', p.last_name, ' (', p.country, ')') SEPARATOR ', ') as players
 			FROM   {$wpdb->prefix}ekc_team t
 			LEFT OUTER JOIN {$wpdb->prefix}ekc_player p
 			ON t.team_id = p.team_id
 			WHERE t.tournament_id = %d
 			      {$sql_filter} 
-			GROUP BY t.team_id, t.name, t.country, t.is_active, t.email, t.phone, t.registration_date, t.registration_order, t.camping_count, t.breakfast_count, t.seeding_score, t.is_on_wait_list
+			GROUP BY t.team_id, t.name, t.country, t.club, t.is_active, t.email, t.phone, t.registration_date, t.registration_order, t.seeding_score, t.is_on_wait_list
 			ORDER BY {$sql_sort}
 			",
 			$tournament_id),
@@ -477,8 +513,8 @@ class Ekc_Database_Access {
 	private function create_team_table_sort_column_sql( $column, $table_alias ) {
 		$sql_alias = $table_alias ? $table_alias . '.' : '';
 		$valid_columns = array(
-		  'name', 'is_active', 'country', 'registration_date', 'registration_order',
-		  'is_on_wait_list', 'camping_count', 'breakfast_count', 'seeding_score'
+		  'name', 'is_active', 'country', 'club', 'registration_date', 'registration_order',
+		  'is_on_wait_list', 'seeding_score'
 		);
 		if ( in_array( $column, $valid_columns, true) ) {
 			return $sql_alias . $column;
@@ -508,15 +544,26 @@ class Ekc_Database_Access {
 		global $wpdb;
 		$results = $wpdb->get_results( $wpdb->prepare(
 			"
-			SELECT  'team_id', 'name', 'country', 'active', 'email', 'phone', 'registration_date', 'order', 'camping', 'breakfast', 'wait_list', 'seeding_score', 'initial_score', 'virtual_rank', 'players'
+			SELECT  'team_id', 'name',
+			        'country', 'club',
+			        'active', 'email', 'phone', 'registration_date', 'order', 
+					'camping', 'breakfast', 
+					'wait_list', 'seeding_score', 'initial_score', 'virtual_rank',
+					'players'
 			UNION ALL
-			SELECT	t.team_id, t.name, LOWER(t.country) as country, case when t.is_active = 1 then 'yes' else 'no' end as is_active, t.email, t.phone, t.registration_date, t.registration_order, t.camping_count, t.breakfast_count, case when t.is_on_wait_list = 1 then 'yes' else 'no' end as is_on_wait_list, t.seeding_score, t.initial_score, t.virtual_rank,
-                               GROUP_CONCAT( CONCAT(p.first_name, ' ', p.last_name, ' (', p.country, ')') SEPARATOR ', ') as players
+			SELECT	t.team_id, t.name, 
+			        LOWER(t.country) as country, t.club, 
+					case when t.is_active = 1 then 'yes' else 'no' end as is_active, 
+					t.email, t.phone, t.registration_date, t.registration_order, 
+					t.camping_count, t.breakfast_count, 
+					case when t.is_on_wait_list = 1 then 'yes' else 'no' end as is_on_wait_list, 
+					t.seeding_score, t.initial_score, t.virtual_rank,
+                    GROUP_CONCAT( CONCAT(p.first_name, ' ', p.last_name, ' (', p.country, ')') SEPARATOR ', ') as players
 			FROM   {$wpdb->prefix}ekc_team t
 			LEFT OUTER JOIN {$wpdb->prefix}ekc_player p
 			ON t.team_id = p.team_id
 			WHERE t.tournament_id = %d
-			GROUP BY t.team_id, t.registration_order, t.name, t.country, t.is_active, t.email, t.phone, t.registration_date, t.camping_count, t.breakfast_count, t.is_on_wait_list, t.seeding_score, t.initial_score, t.virtual_rank
+			GROUP BY t.team_id, t.registration_order, t.name, t.country, t.club, t.is_active, t.email, t.phone, t.registration_date, t.camping_count, t.breakfast_count, t.is_on_wait_list, t.seeding_score, t.initial_score, t.virtual_rank
 			",
 			$tournament_id),
 		ARRAY_A );
@@ -531,20 +578,21 @@ class Ekc_Database_Access {
 			array( 
 				'tournament_id'		=> $team->get_tournament_id(), 
 				'name'			=> $this->truncate_string( $team->get_name(), 500 ),
-				'country'		=> $this->truncate_string( $team->get_country(), 20),
-				'is_active'		=> intval( $team->is_active() ),
+				'country'		=> $this->truncate_string( $team->get_country(), 20 ),
+				'club'          => $this->truncate_string( $team->get_club(), 500 ),
+				'is_active'		=> intval( $team->is_active()),
 				'email'			=> $this->truncate_string( $team->get_email(), 500 ),
 				'phone'			=> $this->truncate_string( $team->get_phone(), 50 ),
 				'registration_date'	=> $team->get_registration_date(),
 				'camping_count'	=> $team->get_camping_count(),
 				'breakfast_count'	=> $team->get_breakfast_count(),
-				'is_on_wait_list'	=> intval( $team->is_on_wait_list() ),
+				'is_on_wait_list'	=> intval( $team->is_on_wait_list()),
 				'registration_order' => $team->get_registration_order(),
 				'seeding_score'	=> $team->get_seeding_score(),
 				'initial_score'	=> $team->get_initial_score(),
 				'virtual_rank'	=> $team->get_virtual_rank(),
 			), 
-			array( '%d', '%s', '%s', '%d', '%s', '%s', '%s', '%d', '%d', '%d', '%f', '%f', '%f', '%d' ) 
+			array( '%d', '%s', '%s', '%s', '%d', '%s', '%s', '%s', '%d', '%d', '%d', '%f', '%f', '%f', '%d' ) 
 		);
 
 		$team_id = $wpdb->insert_id;
@@ -573,7 +621,8 @@ class Ekc_Database_Access {
 			array( 
 				'tournament_id'		=> $team->get_tournament_id(), 
 				'name'			=> $this->truncate_string( $team->get_name(), 500 ),
-				'country'		=> $this->truncate_string( $team->get_country(), 20),
+				'country'		=> $this->truncate_string( $team->get_country(), 20 ),
+				'club'          => $this->truncate_string( $team->get_club(), 500 ),
 				'is_active'		=> intval( $team->is_active() ),
 				'email'			=> $this->truncate_string( $team->get_email(), 500 ),
 				'phone'			=> $this->truncate_string( $team->get_phone(), 50 ),
@@ -586,7 +635,7 @@ class Ekc_Database_Access {
 				'virtual_rank'	=> $team->get_virtual_rank(),
 			),
 			array( 'team_id'		=> $team->get_team_id() ),
-			array( '%d', '%s', '%s', '%d', '%s', '%s', '%d', '%d', '%d', '%f', '%f', '%f', '%d' ),
+			array( '%d', '%s', '%s', '%s', '%d', '%s', '%s', '%d', '%d', '%d', '%f', '%f', '%f', '%d' ),
 			array( '%d' )
 		);
 
@@ -598,11 +647,17 @@ class Ekc_Database_Access {
 		global $wpdb;
 		$row = $wpdb->get_row( $wpdb->prepare( 
 			"
-			SELECT team_id, tournament_id, name, LOWER(country) as country, is_active, email, phone, registration_date, camping_count, breakfast_count, is_on_wait_list, registration_order, seeding_score, initial_score, virtual_rank, shareable_link_id 
+			SELECT team_id, tournament_id, name, 
+			       LOWER(country) as country, club, 
+				   is_active, email, phone, registration_date, 
+				   camping_count, breakfast_count, 
+				   is_on_wait_list, registration_order, 
+				   seeding_score, initial_score, virtual_rank, 
+				   shareable_link_id 
 			FROM   {$wpdb->prefix}ekc_team
 			WHERE  team_id = %d
 			",
-        		$team_id
+        	$team_id
 		) ); 
 
 		if ( ! $row ) {