+++ CHANGELOG.md
@@ -2,9 +2,12 @@ CHANGELOG ekc-tournament
 ========================
 
 ## TODO
-[2] swiss system: allow removal of teams during tournament
 [3] team / tournament: form input/label alignment  
 [3] performance: drop down lists as key / value (integer keys instead of string keys) 
+
+# Version 1.2.14
+[x] Additional attribute on teams: sports club / city (optional column for registration table)
+[x] Removed attributes 'camping count' and 'breakfast count' from teams overview table
 
 # Version 1.2.13
 [x] Fix timer: off by one minute